Streamlit/requirements.txt
# python==3.10.17
pandas==2.2.3
requests==2.32.3
pillow==11.2.1
json5==0.12.0
langchain==0.3.25
langchain-huggingface==0.3.0
langchain-qdrant==0.2.0
langchain-groq==0.3.2
qdrant-client==1.14.3
transformers==4.52.4